// frontend/src/api/models/createReleaseRequest.ts
/**
 * Generated by orval v6.8.1 🍺
 * Do not edit manually.
 * Service Release Manager API
 * OpenAPI spec version: v1
 */
import type { CreateReleaseLocalisedMetadata } from './createReleaseLocalisedMetadata';

export interface CreateReleaseRequest {
	version: string;
	metaData: string;
	localisedMetadataList: CreateReleaseLocalisedMetadata[];
	serviceId: number;
}